// prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Type = {
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE'
} as const

export type Type = (typeof Type)[keyof typeof Type]


export const ExpenseCategory = {
  FOOD: 'FOOD',
  RESTAURANT: 'RESTAURANT',
  SUPERMARKET: 'SUPERMARKET',
  TRANSPORT: 'TRANSPORT',
  FUEL: 'FUEL',
  PUBLIC_TRANSPORT: 'PUBLIC_TRANSPORT',
  VEHICLE_MAINTENANCE: 'VEHICLE_MAINTENANCE',
  HOUSING: 'HOUSING',
  RENT: 'RENT',
  CONDOMINIUM: 'CONDOMINIUM',
  WATER: 'WATER',
  ELECTRICITY: 'ELECTRICITY',
  GAS: 'GAS',
  INTERNET: 'INTERNET',
  PHONE: 'PHONE',
  HEALTH: 'HEALTH',
  PHARMACY: 'PHARMACY',
  MEDICAL: 'MEDICAL',
  DENTIST: 'DENTIST',
  EDUCATION: 'EDUCATION',
  COURSES: 'COURSES',
  BOOKS: 'BOOKS',
  ENTERTAINMENT: 'ENTERTAINMENT',
  STREAMING: 'STREAMING',
  GAMES: 'GAMES',
  MOVIES: 'MOVIES',
  SHOPPING: 'SHOPPING',
  CLOTHING: 'CLOTHING',
  ELECTRONICS: 'ELECTRONICS',
  TRAVEL: 'TRAVEL',
  INSURANCE: 'INSURANCE',
  TAXES: 'TAXES',
  PETS: 'PETS',
  OTHER: 'OTHER'
} as const

export type ExpenseCategory = (typeof ExpenseCategory)[keyof typeof ExpenseCategory]


export const IncomeCategory = {
  SALARY: 'SALARY',
  BONUS: 'BONUS',
  FREELANCE: 'FREELANCE',
  COMMISSION: 'COMMISSION',
  INVESTMENT: 'INVESTMENT',
  DIVIDENDS: 'DIVIDENDS',
  RENTAL: 'RENTAL',
  REFUND: 'REFUND',
  GIFT: 'GIFT',
  SALE: 'SALE',
  OTHER: 'OTHER'
} as const

export type IncomeCategory = (typeof IncomeCategory)[keyof typeof IncomeCategory]